[util] Move shape options into shape-options.hh
