+++ util/Makefile.sources
@@ -7,6 +7,7 @@ HB_VIEW_sources = \
 	options.cc \
 	options.hh \
 	shape-consumer.hh \
+	shape-options.hh \
 	view-cairo.hh \
 	view-options.hh \
 	$(NULL)
@@ -17,6 +18,7 @@ HB_SHAPE_sources = \
 	options.cc \
 	options.hh \
 	shape-consumer.hh \
+	shape-options.hh \
 	$(NULL)
 
 HB_OT_SHAPE_CLOSURE_sources = \
